[scip] refactor for clearer structure
