+++ codeminer/scip_interface/scip_decode.py
@@ -1,8 +1,8 @@
 """
 Unified SCIP decoder that supports multiple languages (C++, Rust, TypeScript, Python).
 
-This module provides a single SCIPDecoder class that can decode SCIP indexes for multiple
-languages by selecting the appropriate language-specific decoder based on the 'language' parameter.
+This module provides a single SCIPGraphDecoder class that routes to language-specific
+decoder implementations based on the 'language' parameter.
 
 Supported languages:
     - 'cpp', 'c++', 'c': C/C++ projects
@@ -12,42 +12,31 @@ Supported languages:
 
 Example:
     # C++ decoder
-    decoder = SCIPDecoder(index_file_path="path/to/index.decoded", language="cpp")
+    decoder = SCIPGraphDecoder(index_file_path="path/to/index.decoded", language="cpp")
     graph = decoder.decode()
 
     # Rust decoder
-    decoder = SCIPDecoder(index_file_path="path/to/index.decoded", language="rust")
+    decoder = SCIPGraphDecoder(index_file_path="path/to/index.decoded", language="rust")
     graph = decoder.decode()
 
-    # Python decoder (original behavior)
-    decoder = SCIPDecoder(index_file_path="path/to/index.decoded", language="python")
+    # Python decoder (default)
+    decoder = SCIPGraphDecoder(index_file_path="path/to/index.decoded", language="python")
     # Or omit language parameter for backward compatibility
-    decoder = SCIPDecoder(index_file_path="path/to/index.decoded")
+    decoder = SCIPGraphDecoder(index_file_path="path/to/index.decoded")
 """
-import re
-from pathlib import Path
 from typing import Optional
 
-from ..graph.code_graph import CodeGraph
-from ..log_utils import get_logger, register_scip_logger
-from ..types import (
-    EDGE_TYPE_CONTAIN,
-    EDGE_TYPE_REFERENCE,
-    NODE_TYPE_CLASS,
-    NODE_TYPE_FIELD,
-    NODE_TYPE_FUNCTION,
-    NODE_TYPE_METHOD,
-    ROOT_NODE,
-)
+from ..log_utils import get_logger
+
+logger = get_logger("scip_decode")
 
 
-class SCIPDecoder:
+class SCIPGraphDecoder:
     """
-    Unified SCIP decoder that supports multiple languages.
+    Unified SCIP decoder that delegates to language-specific implementations.
 
-    This class can work in two modes:
-    1. Language-agnostic mode (default, backward compatible): Acts as Python SCIP decoder
-    2. Multi-language mode: Delegates to language-specific decoders based on 'language' parameter
+    This class acts as a router: it selects the appropriate language-specific
+    decoder based on the 'language' parameter and delegates all operations to it.
     """
 
     # Language aliases for flexibility
@@ -88,7 +77,6 @@ class SCIPDecoder:
         # Normalize language name (None means Python for backward compatibility)
         if language is None:
             self.language = 'python'
-            self._delegate = None  # Use built-in Python decoder
         else:
             language_lower = language.lower()
             if language_lower not in self.LANGUAGE_ALIASES:
@@ -99,24 +87,14 @@ class SCIPDecoder:
                 )
             self.language = self.LANGUAGE_ALIASES[language_lower]
 
-            # Create delegate for non-Python languages
-            if self.language != 'python':
-                self._delegate = self._create_language_decoder(
-                    index_file_path=index_file_path,
-                    project_root=project_root,
-                )
-                # Copy code_graph from delegate
-                self.code_graph = self._delegate.code_graph
-                return
-            else:
-                self._delegate = None
+        # Create the language-specific delegate
+        self._delegate = self._create_language_decoder(
+            index_file_path=index_file_path,
+            project_root=project_root,
+        )
 
-        # Python-specific initialization (original behavior)
-        self.code_graph = CodeGraph(project_root)
-        self.indexed_directories = set()
-        self.logger = get_logger(__name__)
-        # Register this module for SCIP debug logging
-        register_scip_logger(__name__)
+        # Expose delegate's code_graph for backward compatibility
+        self.code_graph = self._delegate.code_graph
 
     def _create_language_decoder(
         self,
@@ -125,10 +103,6 @@ class SCIPDecoder:
     ):
         """
         Create the appropriate language-specific decoder.
-
-        Args:
-            index_file_path: Path to the decoded SCIP index file
-            project_root: Root directory of the project
 
         Returns:
             Language-specific decoder instance
@@ -154,355 +128,22 @@ class SCIPDecoder:
                 project_root=project_root,
             )
 
+        elif self.language == 'python':
+            from .scip_decode_python import SCIPPythonGraphDecoder
+            return SCIPPythonGraphDecoder(
+                index_file_path=index_file_path,
+                project_root=project_root,
+            )
+
         else:
             raise ValueError(f"No decoder implementation for language: {self.language}")
 
+    # ── Delegated methods ──────────────────────────────────────────────
+
     def decode(self):
-        """
-        Decode the SCIP index into a CodeGraph.
-
-        Returns:
-            CodeGraph: The decoded graph
-        """
-        # Delegate to language-specific decoder if available
-        if self._delegate is not None:
-            return self._delegate.decode()
-
-        # Python-specific implementation (original behavior)
-        self.logger.info(f"Starting SCIP Python decode from {self.index_file_path}")
-        try:
-            with open(self.index_file_path, "r") as f:
-                content = f.read()
-        except Exception as e:
-            self.logger.error(f"Error reading SCIP index file: {e}")
-            raise
-
-        # Parse documents
-        document_blocks = re.findall(
-            r"documents\s*{(.*?)(?=documents\s*{|$)", content, re.DOTALL
-        )
-
-        # Add the root node to the graph
-        self.code_graph.add_root_node(ROOT_NODE)
-
-        # Process all documents
-        for document in document_blocks:
-            self._process_document(document)
-
-        return self.code_graph
+        """Decode the SCIP index into a CodeGraph."""
+        return self._delegate.decode()
 
     def save_graph(self, output_path: str):
-        """
-        Save the decoded graph to a file.
-
-        Args:
-            output_path: Path to save the graph
-        """
-        # Delegate to language-specific decoder if available
-        if self._delegate is not None:
-            return self._delegate.save_graph(output_path)
-
-        # Python-specific implementation
-        self.code_graph.save_graph(output_path)
-
-
-# Backward compatibility alias
-class SCIPGraphDecoder(SCIPDecoder):
-    """
-    Backward compatibility alias for SCIPDecoder.
-
-    This class exists to maintain compatibility with existing code that uses SCIPGraphDecoder.
-    New code should use SCIPDecoder instead.
-    """
-    def __init__(self, index_file_path, project_root=None):
-        super().__init__(index_file_path=index_file_path, project_root=project_root, language='python')
-
-    def decode(self):
-        self.logger.info(f"Starting SCIP decode from {self.index_file_path}")
-        try:
-            with open(self.index_file_path, "r") as f:
-                content = f.read()
-        except Exception as e:
-            self.logger.error(f"Error reading SCIP index file: {e}")
-            raise
-
-        # Parse documents
-        document_blocks = re.findall(
-            r"documents\s*{(.*?)(?=documents\s*{|$)", content, re.DOTALL
-        )
-
-        # Add the root node to the graph
-        self.code_graph.add_root_node(ROOT_NODE)
-
-        # Process all documents
-        for document in document_blocks:
-            self._process_document(document)
-
-        return self.code_graph
-
-    def _process_document(self, document_text):
-        # Extract file path
-        file_match = re.search(r'relative_path:\s*"([^"]+)"', document_text)
-        if not file_match:
-            return
-
-        file_path = file_match.group(1)
-
-        # Iteratively Extract the directory path from the file path
-        dir_path = Path(file_path).parent
-        while dir_path != dir_path.parent:  # Stop at the root directory
-            dir_path_str = str(dir_path)
-            if dir_path_str not in self.indexed_directories:
-                # Add directory node if not already indexed
-                self.code_graph.add_directory_node(dir_path_str)
-                self.indexed_directories.add(dir_path_str)
-                # Add containment edge from parent directory to this directory
-                self.code_graph._add_edge(
-                    str(dir_path.parent), dir_path_str, EDGE_TYPE_CONTAIN
-                )
-            dir_path = dir_path.parent
-
-        # Add file node
-        self.code_graph.add_file_node(file_path)
-
-        # Add file containment edge
-        self.code_graph._add_edge(
-            str(Path(file_path).parent), file_path, EDGE_TYPE_CONTAIN
-        )
-
-        # Process occurrences
-        occurrences = re.findall(r"occurrences\s*{(.*?)}", document_text, re.DOTALL)
-        for occurrence in occurrences:
-            self._process_occurrence(occurrence)
-
-    def _process_occurrence(self, occurrence_text):
-        # Skip stdlib symbols
-        if "python-stdlib" in occurrence_text:
-            return
-
-        # Extract ranges
-        ranges = re.findall(r"range:\s*(\d+)", occurrence_text)
-        if len(ranges) < 3:
-            return
-
-        line = int(ranges[0])
-
-        # Extract symbol
-        symbol_match = re.search(r'symbol:\s*"([^"]+)"', occurrence_text)
-        if not symbol_match:
-            return
-
-        symbol = symbol_match.group(1)
-
-        # Skip local symbols (scip represents them as 'local <id>')
-        if symbol.startswith("local "):
-            return
-
-        # Extract symbol_roles
-        symbol_roles_match = re.search(r"symbol_roles:\s*(\d+)", occurrence_text)
-        if not symbol_roles_match:
-            return
-
-        symbol_roles = int(symbol_roles_match.group(1))
-
-        # Extract enclosing range if available
-        enclosing_ranges = re.findall(r"enclosing_range:\s*(\d+)", occurrence_text)
-
-        # Process the symbol
-        self._process_symbol(symbol, line, symbol_roles, enclosing_ranges)
-
-    def _unify_symbol_name(self, symbol):
-        """
-        Unify symbol names to a consistent format.
-
-        Examples:
-        - src.calculator`/Calculator# -> src/calculator.py:Calculator
-        - src.calculator`/Calculator#add(). -> src/calculator.py:Calculator.add()
-        - src.utils.helpers`/validate_input(). -> src/utils/helpers.py:validate_input()
-        - src.calculator`/Calculator#history. -> src/calculator.py:Calculator.history (assuming history is a field)
-
-        Args:
-            symbol: Original symbol name
-
-        Returns:
-            Unified symbol name
-        """
-        # Remove backticks
-        clean_symbol = symbol.replace("`", "")
-
-        # Replace dots with slashes in module path and use colon as separator
-        if "/" in clean_symbol:
-            parts = clean_symbol.split("/", 1)
-            module_path = parts[0].replace(".", "/") + ".py"  # Add .py suffix
-            if len(parts) > 1:
-                symbol_part = parts[1]
-
-                # Handle class and method patterns
-                if "#" in symbol_part:
-                    # Split on # to separate class from method
-                    class_method_parts = symbol_part.split("#", 1)
-                    class_name = class_method_parts[0]
-
-                    if len(class_method_parts) > 1 and class_method_parts[1]:
-                        # Has method after #
-                        method_part = class_method_parts[1].rstrip(".")
-                        unified = f"{module_path}:{class_name}.{method_part}"
-                    else:
-                        # Just class (ends with #)
-                        unified = f"{module_path}:{class_name}"
-                else:
-                    # Function (no # symbol)
-                    func_name = symbol_part.rstrip(".")
-                    unified = f"{module_path}:{func_name}"
-            else:
-                unified = module_path
-        else:
-            # No module path separator, use as-is but clean
-            unified = clean_symbol.rstrip(".")
-
-        return unified
-
-    def _classify_symbol_type(self, unified_symbol, original_symbol=None):
-        """
-        Classify symbol type based on unified symbol format.
-
-        Args:
-            unified_symbol: Unified symbol name
-            original_symbol: Original symbol name (for additional context)
-
-        Returns:
-            Symbol type: NODE_TYPE_CLASS, NODE_TYPE_METHOD, NODE_TYPE_FIELD, or NODE_TYPE_FUNCTION
-        """
-
-        if ":" in unified_symbol:
-            symbol_part = unified_symbol.split(":", 1)[1]
-            if "." in symbol_part:
-                # Has a dot - could be method or field
-                # Check if the original symbol had parentheses (indicating method)
-                has_parentheses = False
-                if original_symbol:
-                    has_parentheses = "()" in original_symbol or "(" in original_symbol
-
-                # If original had parentheses, it's a method; otherwise it's a field
-                if has_parentheses:
-                    return NODE_TYPE_METHOD
-                else:
-                    return NODE_TYPE_FIELD
-            else:
-                # Could be class or function - check if it looks like a class
-                # Classes typically start with capital letter
-                if symbol_part and symbol_part[0].isupper():
-                    return NODE_TYPE_CLASS
-                else:
-                    return NODE_TYPE_FUNCTION
-        else:
-            return NODE_TYPE_FUNCTION
-
-    def _process_symbol(self, symbol, line, symbol_roles, enclosing_ranges):
-        self.logger.scip_debug(
-            f"Processing symbol: {symbol} at line {line}, roles: {symbol_roles}"
-        )
-
-        # Skip function arguments (symbols ending with .(xxx))
-        if re.search(r"\.\([^)]+\)$", symbol):
-            return
-
-        # Exit scopes that have ended based on current line
-        try:
-            self.logger.scip_debug(
-                f"Scope stack before exit: {[list(s.keys())[0] for s in self.code_graph.scope_stack]}"
-            )
-            self.code_graph.exit_scopes_by_line(line)
-            self.logger.scip_debug(
-                f"Scope stack after exit: {[list(s.keys())[0] for s in self.code_graph.scope_stack]}"
-            )
-        except Exception as e:
-            self.logger.error(f"Error exiting scopes at line {line}: {e}")
-            raise
-
-        # Parse the symbol
-        match = re.search(r"`?([^`]+)`?/([^.]+)(?:\.|\(|#)", symbol)
-        if not match:
-            return
-
-        module_path = match.group(1)
-
-        # Clean up the symbol by simply splitting on spaces and taking the last part
-        # For example: "scip-python python HttpieCliRepo 5b604c37c6c67e18e7c3e9aee6c88a8c22b98345 extras.profiling.benchmarks/QuietSimpleHTTPServer#log_message()."
-        # Will become: "extras.profiling.benchmarks/QuietSimpleHTTPServer#log_message()."
-        cleaned_symbol = symbol.split(" ")[-1]
-        cleaned_symbol = re.sub(r"`", "", cleaned_symbol)
-
-        # Unify symbol name format
-        unified_symbol = self._unify_symbol_name(cleaned_symbol)
-
-        # Classify symbol type (pass both original and unified for context)
-        symbol_type = self._classify_symbol_type(unified_symbol, cleaned_symbol)
-
-        # Handle __init__ symbols - convert to file reference
-        if "/__init__" in unified_symbol:
-            # Extract the module path and use it as the target
-            module_match = re.search(r"(.+)/(?:__init__)", unified_symbol)
-            if module_match:
-                module_path = module_match.group(1)
-                file_path = module_path.replace(".", "/") + ".py"
-
-                # If this is a reference, point to the file instead
-                if symbol_roles == 8:
-                    self.code_graph._add_edge(
-                        self.code_graph.current_scope, file_path, EDGE_TYPE_REFERENCE
-                    )
-                return
-
-        # Update current scope if this is a definition with enclosing range
-        if symbol_roles == 1 and enclosing_ranges and len(enclosing_ranges) >= 4:
-            scope_start_line = int(enclosing_ranges[0])
-            scope_end_line = int(enclosing_ranges[2])
-
-            # Add symbol node with scope range
-            self.code_graph.add_symbol_node(
-                unified_symbol, line, scope_start_line, scope_end_line, symbol_type
-            )
-
-            # Add containment edge
-            self.logger.scip_debug(
-                f"Adding containment edge for {unified_symbol}, current scope: {self.code_graph.current_scope}"
-            )
-            self.code_graph.add_containment_edge(unified_symbol)
-
-            # Update current scope for classes and functions with enclosing ranges
-            # Now with proper scope exit handling, functions can safely become scopes
-            if symbol_type in [NODE_TYPE_CLASS, NODE_TYPE_FUNCTION, NODE_TYPE_METHOD]:
-                try:
-                    self.logger.scip_debug(
-                        f"Updating scope to {unified_symbol} [{scope_start_line}-{scope_end_line}]"
-                    )
-                    self.code_graph.update_current_scope(
-                        unified_symbol, scope_start_line, scope_end_line
-                    )
-                except Exception as e:
-                    self.logger.error(f"Error updating scope for {unified_symbol}: {e}")
-                    raise
-
-        # Handle definition (symbol_roles == 1) with no enclosing range
-        elif symbol_roles == 1:
-            self.logger.scip_debug(
-                f"Adding symbol without enclosing range: {unified_symbol}, current scope: {self.code_graph.current_scope}"
-            )
-            self.code_graph.add_symbol_node(
-                unified_symbol, line, symbol_type=symbol_type
-            )
-
-            # Add 'contain' edge from current scope to symbol
-            self.code_graph._add_edge(
-                self.code_graph.current_scope, unified_symbol, EDGE_TYPE_CONTAIN
-            )
-
-        # Handle reference (symbol_roles == 8)
-        elif symbol_roles == 8:
-            self.code_graph.add_symbol_reference(
-                unified_symbol, module_path, symbol_type
-            )
-
-    def save_graph(self, output_path):
-        self.code_graph.save_graph(output_path)
+        """Save the decoded graph to a file."""
+        return self._delegate.save_graph(output_path)
